+++ frontend/super-frontend/source/src/javascripts/application.js
@@ -1,8 +1,11 @@
+import "details-element-polyfill";
+
 import RailsUjs from "@rails/ujs";
 import { Application, Controller as StimulusController } from "stimulus";
 import ApplyTemplateController from "./apply_template_controller";
 import CleanFilterParamController from "./clean_filter_param_controller";
 import CleanFilterParamsController from "./clean_filter_params_controller";
+import ClickOutsideToCloseController from "./click_outside_to_close_controller";
 import DeleteController from "./delete_controller";
 import FlatpickrController from "./flatpickr_controller";
 import TogglePendingDestructionController from "./toggle_pending_destruction_controller";
@@ -13,6 +16,10 @@ StimulusApplication.register("clean-filter-param", CleanFilterParamController);
 StimulusApplication.register(
   "clean-filter-params",
   CleanFilterParamsController
+);
+StimulusApplication.register(
+  "click-outside-to-close",
+  ClickOutsideToCloseController
 );
 StimulusApplication.register("delete", DeleteController);
 StimulusApplication.register("flatpickr", FlatpickrController);