+++ lib/super/link.rb
@@ -22,57 +22,57 @@ module Super
 
     def self.registry
       @registry ||= {
-        new: new(
+        new: LinkBuilder.new(
           "New",
           -> (params:) {
-            Rails.application.routes.url_for(
+            {
               controller: params[:controller],
               action: :new,
               only_path: true
-            )
+            }
           }
         ),
-        index: new(
+        index: LinkBuilder.new(
           "Index",
           -> (params:) {
-            Rails.application.routes.url_for(
+            {
               controller: params[:controller],
               action: :index,
               only_path: true
-            )
+            }
           }
         ),
-        show: new(
+        show: LinkBuilder.new(
           "View",
           -> (record:, params:) {
-            Rails.application.routes.url_for(
+            {
               controller: params[:controller],
               action: :show,
               id: record,
               only_path: true
-            )
+            }
           }
         ),
-        edit: new(
+        edit: LinkBuilder.new(
           "Edit",
           -> (record:, params:) {
-            Rails.application.routes.url_for(
+            {
               controller: params[:controller],
               action: :edit,
               id: record,
               only_path: true
-            )
+            }
           }
         ),
-        destroy: new(
+        destroy: LinkBuilder.new(
           "Delete",
           -> (record:, params:) {
-            Rails.application.routes.url_for(
+            {
               controller: params[:controller],
               action: :destroy,
               id: record,
               only_path: true
-            )
+            }
           },
           method: :delete,
           data: { confirm: "Really delete?" }
@@ -91,27 +91,33 @@ module Super
       @options = options
     end
 
-    def to_s(default_options: nil, **proc_arguments)
+    attr_reader :text
+    attr_reader :options
+
+    def href
+      if @href.is_a?(String)
+        return @href
+      end
+
+      if @href.is_a?(Hash)
+        @href = Rails.application.routes.url_for(**@href)
+        return @href
+      end
+
+      @href = Super::Compatability.polymorphic_path_container.polymorphic_path(@href)
+    end
+
+    def to_s(default_options: nil)
       default_options ||= {}
       ActionController::Base.helpers.link_to(
-        value(text, proc_arguments),
-        value(href, proc_arguments),
-        default_options.deep_merge(value(options, proc_arguments))
+        text,
+        href,
+        default_options.deep_merge(options)
       )
     end
 
-    private
-
-    attr_reader :text
-    attr_reader :href
-    attr_reader :options
-
-    def value(proc_or_value, proc_arguments)
-      if proc_or_value.kind_of?(Proc)
-        proc_or_value.call(**proc_arguments)
-      else
-        proc_or_value
-      end
+    def ==(other)
+      self.class == other.class && text == other.text && href == other.href && options == other.options
     end
   end
 end